Revert "[ObjC]fix 64 to 32 bit clang conversion warning in src/core/lib (#29877)"
This reverts commit eb0a5918116fa7c600ab1e2fb47782a5c55c55b9.

+++ include/grpcpp/test/channel_test_peer.h
@@ -31,7 +31,7 @@ class ChannelTestPeer {
 
   /// Provide the gRPC Core channel
   grpc_channel* channel() const { return channel_->c_channel_; }
-  size_t registered_calls() const;
+  int registered_calls() const;
   int registration_attempts() const;
 
  private:
